+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -538,18 +538,101 @@ func (a *AmazonRdsForSQLServerLinkedService) GetLinkedService() *LinkedService {
 
 // AmazonRdsForSQLServerLinkedServiceTypeProperties - Amazon Rds for SQL Server linked service properties.
 type AmazonRdsForSQLServerLinkedServiceTypeProperties struct {
-	// REQUIRED; The connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
-	ConnectionString any
-
 	// Sql always encrypted properties.
 	AlwaysEncryptedSettings *SQLAlwaysEncryptedProperties
+
+	// The application workload type when connecting to a server, used by recommended version. Possible values are ReadOnly and
+	// ReadWrite. Type: string.
+	ApplicationIntent any
+
+	// The type used for authentication. Type: string.
+	AuthenticationType *AmazonRdsForSQLAuthenticationType
+
+	// The default wait time (in seconds) before terminating the attempt to execute a command and generating an error, used by
+	// recommended version. Type: integer.
+	CommandTimeout any
+
+	// The number of re-connections attempted after identifying that there was an idle connection failure, used by recommended
+	// version. This must be an integer between 0 and 255. Type: integer.
+	ConnectRetryCount any
+
+	// The amount of time (in seconds) between each re-connection attempt after identifying that there was an idle connection
+	// failure, used by recommended version. This must be an integer between 1 and 60.
+	// Type: integer.
+	ConnectRetryInterval any
+
+	// The length of time (in seconds) to wait for a connection to the server before terminating the attempt and generating an
+	// error, used by recommended version. Type: integer.
+	ConnectTimeout any
+
+	// The connection string, used by legacy version. Type: string, SecureString or AzureKeyVaultSecretReference. Type: string,
+	// SecureString or AzureKeyVaultSecretReference.
+	ConnectionString any
+
+	// The name of the database, used by recommended version. Type: string.
+	Database any
+
+	// Indicate whether TLS encryption is required for all data sent between the client and server, used by recommended version.
+	// Type: string.
+	Encrypt any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// The name or address of the partner server to connect to if the primary server is down, used by recommended version. Type:
+	// string.
+	FailoverPartner any
+
+	// The host name to use when validating the server certificate for the connection. When not specified, the server name from
+	// the Data Source is used for certificate validation, used by recommended
+	// version. Type: string.
+	HostNameInCertificate any
+
+	// Indicate whether User ID and Password are specified in the connection (when false) or whether the current Windows account
+	// credentials are used for authentication (when true), used by recommended
+	// version. Type: Boolean.
+	IntegratedSecurity any
+
+	// The minimum time, in seconds, for the connection to live in the connection pool before being destroyed, used by recommended
+	// version. Type: integer.
+	LoadBalanceTimeout any
+
+	// The maximum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MaxPoolSize any
+
+	// The minimum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MinPoolSize any
+
+	// If your application is connecting to an AlwaysOn availability group (AG) on different subnets, setting MultiSubnetFailover=true
+	// provides faster detection of and connection to the (currently) active
+	// server, used by recommended version. Type: Boolean.
+	MultiSubnetFailover any
+
+	// When true, an application can maintain multiple active result sets (MARS). When false, an application must process or cancel
+	// all result sets from one batch before it can execute any other batch on
+	// that connection, used by recommended version. Type: Boolean.
+	MultipleActiveResultSets any
+
+	// The size in bytes of the network packets used to communicate with an instance of server, used by recommended version. Type:
+	// integer.
+	PacketSize any
+
 	// The on-premises Windows authentication password.
 	Password SecretBaseClassification
+
+	// Indicate whether the connection will be pooled or explicitly opened every time that the connection is requested, used by
+	// recommended version. Type: Boolean.
+	Pooling any
+
+	// The name or network address of the instance of SQL Server to which to connect, used by recommended version. Type: string.
+	Server any
+
+	// Indicate whether the channel will be encrypted while bypassing walking the certificate chain to validate trust, used by
+	// recommended version. Type: Boolean.
+	TrustServerCertificate any
 
 	// The on-premises Windows authentication user name. Type: string (or Expression with resultType string).
 	UserName any
@@ -3689,7 +3772,7 @@ type AzureFunctionActivityTypeProperties struct {
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
 	// string (or Expression with resultType string).
-	Headers map[string]*string
+	Headers map[string]any
 }
 
 // AzureFunctionLinkedService - Azure Function linked service.
@@ -4890,24 +4973,113 @@ func (a *AzureSQLDWLinkedService) GetLinkedService() *LinkedService {
 
 // AzureSQLDWLinkedServiceTypeProperties - Azure SQL Data Warehouse linked service properties.
 type AzureSQLDWLinkedServiceTypeProperties struct {
-	// REQUIRED; The connection string. Type: string, SecureString or AzureKeyVaultSecretReference. Type: string, SecureString
-	// or AzureKeyVaultSecretReference.
-	ConnectionString any
+	// The application workload type when connecting to a server, used by recommended version. Possible values are ReadOnly and
+	// ReadWrite. Type: string.
+	ApplicationIntent any
+
+	// The type used for authentication. Type: string.
+	AuthenticationType *AzureSQLDWAuthenticationType
 
 	// Indicates the azure cloud type of the service principle auth. Allowed values are AzurePublic, AzureChina, AzureUsGovernment,
 	// AzureGermany. Default value is the data factory regions’ cloud type. Type:
 	// string (or Expression with resultType string).
 	AzureCloudType any
 
+	// The default wait time (in seconds) before terminating the attempt to execute a command and generating an error, used by
+	// recommended version. Type: integer.
+	CommandTimeout any
+
+	// The number of re-connections attempted after identifying that there was an idle connection failure, used by recommended
+	// version. This must be an integer between 0 and 255. Type: integer.
+	ConnectRetryCount any
+
+	// The amount of time (in seconds) between each re-connection attempt after identifying that there was an idle connection
+	// failure, used by recommended version. This must be an integer between 1 and 60.
+	// Type: integer.
+	ConnectRetryInterval any
+
+	// The length of time (in seconds) to wait for a connection to the server before terminating the attempt and generating an
+	// error, used by recommended version. Type: integer.
+	ConnectTimeout any
+
+	// The connection string, used by legacy version. Type: string, SecureString or AzureKeyVaultSecretReference. Type: string,
+	// SecureString or AzureKeyVaultSecretReference.
+	ConnectionString any
+
 	// The credential reference containing authentication information.
 	Credential *CredentialReference
+
+	// The name of the database, used by recommended version. Type: string.
+	Database any
+
+	// Indicate whether TLS encryption is required for all data sent between the client and server, used by recommended version.
+	// Type: string.
+	Encrypt any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// The name or address of the partner server to connect to if the primary server is down, used by recommended version. Type:
+	// string.
+	FailoverPartner any
+
+	// The host name to use when validating the server certificate for the connection. When not specified, the server name from
+	// the Data Source is used for certificate validation, used by recommended
+	// version. Type: string.
+	HostNameInCertificate any
+
+	// Indicate whether User ID and Password are specified in the connection (when false) or whether the current Windows account
+	// credentials are used for authentication (when true), used by recommended
+	// version. Type: Boolean.
+	IntegratedSecurity any
+
+	// The minimum time, in seconds, for the connection to live in the connection pool before being destroyed, used by recommended
+	// version. Type: integer.
+	LoadBalanceTimeout any
+
+	// The maximum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MaxPoolSize any
+
+	// The minimum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MinPoolSize any
+
+	// If your application is connecting to an AlwaysOn availability group (AG) on different subnets, setting MultiSubnetFailover=true
+	// provides faster detection of and connection to the (currently) active
+	// server, used by recommended version. Type: Boolean.
+	MultiSubnetFailover any
+
+	// When true, an application can maintain multiple active result sets (MARS). When false, an application must process or cancel
+	// all result sets from one batch before it can execute any other batch on
+	// that connection, used by recommended version. Type: Boolean.
+	MultipleActiveResultSets any
+
+	// The size in bytes of the network packets used to communicate with an instance of server, used by recommended version. Type:
+	// integer.
+	PacketSize any
+
 	// The Azure key vault secret reference of password in connection string.
 	Password *AzureKeyVaultSecretReference
+
+	// Indicate whether the connection will be pooled or explicitly opened every time that the connection is requested, used by
+	// recommended version. Type: Boolean.
+	Pooling any
+
+	// The name or network address of the instance of SQL Server to which to connect, used by recommended version. Type: string.
+	Server any
+
+	// The credential of the service principal object in Azure Active Directory. If servicePrincipalCredentialType is 'ServicePrincipalKey',
+	// servicePrincipalCredential can be SecureString or
+	// AzureKeyVaultSecretReference. If servicePrincipalCredentialType is 'ServicePrincipalCert', servicePrincipalCredential can
+	// only be AzureKeyVaultSecretReference.
+	ServicePrincipalCredential SecretBaseClassification
+
+	// The service principal credential type to use in Server-To-Server authentication. 'ServicePrincipalKey' for key/secret,
+	// 'ServicePrincipalCert' for certificate. Type: string (or Expression with
+	// resultType string).
+	ServicePrincipalCredentialType any
 
 	// The ID of the service principal used to authenticate against Azure SQL Data Warehouse. Type: string (or Expression with
 	// resultType string).
@@ -4918,6 +5090,13 @@ type AzureSQLDWLinkedServiceTypeProperties struct {
 
 	// The name or ID of the tenant to which the service principal belongs. Type: string (or Expression with resultType string).
 	Tenant any
+
+	// Indicate whether the channel will be encrypted while bypassing walking the certificate chain to validate trust, used by
+	// recommended version. Type: Boolean.
+	TrustServerCertificate any
+
+	// The user name to be used when connecting to server. Type: string.
+	UserName any
 }
 
 // AzureSQLDWTableDataset - The Azure SQL Data Warehouse dataset.
@@ -5019,26 +5198,116 @@ func (a *AzureSQLDatabaseLinkedService) GetLinkedService() *LinkedService {
 
 // AzureSQLDatabaseLinkedServiceTypeProperties - Azure SQL Database linked service properties.
 type AzureSQLDatabaseLinkedServiceTypeProperties struct {
-	// REQUIRED; The connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
-	ConnectionString any
-
 	// Sql always encrypted properties.
 	AlwaysEncryptedSettings *SQLAlwaysEncryptedProperties
+
+	// The application workload type when connecting to a server, used by recommended version. Possible values are ReadOnly and
+	// ReadWrite. Type: string.
+	ApplicationIntent any
+
+	// The type used for authentication. Type: string.
+	AuthenticationType *AzureSQLDatabaseAuthenticationType
 
 	// Indicates the azure cloud type of the service principle auth. Allowed values are AzurePublic, AzureChina, AzureUsGovernment,
 	// AzureGermany. Default value is the data factory regions’ cloud type. Type:
 	// string (or Expression with resultType string).
 	AzureCloudType any
 
+	// The default wait time (in seconds) before terminating the attempt to execute a command and generating an error, used by
+	// recommended version. Type: integer.
+	CommandTimeout any
+
+	// The number of re-connections attempted after identifying that there was an idle connection failure, used by recommended
+	// version. This must be an integer between 0 and 255. Type: integer.
+	ConnectRetryCount any
+
+	// The amount of time (in seconds) between each re-connection attempt after identifying that there was an idle connection
+	// failure, used by recommended version. This must be an integer between 1 and 60.
+	// Type: integer.
+	ConnectRetryInterval any
+
+	// The length of time (in seconds) to wait for a connection to the server before terminating the attempt and generating an
+	// error, used by recommended version. Type: integer.
+	ConnectTimeout any
+
+	// The connection string, used by legacy version. Type: string, SecureString or AzureKeyVaultSecretReference. Type: string,
+	// SecureString or AzureKeyVaultSecretReference.
+	ConnectionString any
+
 	// The credential reference containing authentication information.
 	Credential *CredentialReference
+
+	// The name of the database, used by recommended version. Type: string.
+	Database any
+
+	// Indicate whether TLS encryption is required for all data sent between the client and server, used by recommended version.
+	// Type: string.
+	Encrypt any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// The name or address of the partner server to connect to if the primary server is down, used by recommended version. Type:
+	// string.
+	FailoverPartner any
+
+	// The host name to use when validating the server certificate for the connection. When not specified, the server name from
+	// the Data Source is used for certificate validation, used by recommended
+	// version. Type: string.
+	HostNameInCertificate any
+
+	// Indicate whether User ID and Password are specified in the connection (when false) or whether the current Windows account
+	// credentials are used for authentication (when true), used by recommended
+	// version. Type: Boolean.
+	IntegratedSecurity any
+
+	// The minimum time, in seconds, for the connection to live in the connection pool before being destroyed, used by recommended
+	// version. Type: integer.
+	LoadBalanceTimeout any
+
+	// The maximum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MaxPoolSize any
+
+	// The minimum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MinPoolSize any
+
+	// If your application is connecting to an AlwaysOn availability group (AG) on different subnets, setting MultiSubnetFailover=true
+	// provides faster detection of and connection to the (currently) active
+	// server, used by recommended version. Type: Boolean.
+	MultiSubnetFailover any
+
+	// When true, an application can maintain multiple active result sets (MARS). When false, an application must process or cancel
+	// all result sets from one batch before it can execute any other batch on
+	// that connection, used by recommended version. Type: Boolean.
+	MultipleActiveResultSets any
+
+	// The size in bytes of the network packets used to communicate with an instance of server, used by recommended version. Type:
+	// integer.
+	PacketSize any
+
 	// The Azure key vault secret reference of password in connection string.
 	Password *AzureKeyVaultSecretReference
+
+	// Indicate whether the connection will be pooled or explicitly opened every time that the connection is requested, used by
+	// recommended version. Type: Boolean.
+	Pooling any
+
+	// The name or network address of the instance of SQL Server to which to connect, used by recommended version. Type: string.
+	Server any
+
+	// The credential of the service principal object in Azure Active Directory. If servicePrincipalCredentialType is 'ServicePrincipalKey',
+	// servicePrincipalCredential can be SecureString or
+	// AzureKeyVaultSecretReference. If servicePrincipalCredentialType is 'ServicePrincipalCert', servicePrincipalCredential can
+	// only be AzureKeyVaultSecretReference.
+	ServicePrincipalCredential SecretBaseClassification
+
+	// The service principal credential type to use in Server-To-Server authentication. 'ServicePrincipalKey' for key/secret,
+	// 'ServicePrincipalCert' for certificate. Type: string (or Expression with
+	// resultType string).
+	ServicePrincipalCredentialType any
 
 	// The ID of the service principal used to authenticate against Azure SQL Database. Type: string (or Expression with resultType
 	// string).
@@ -5049,6 +5318,13 @@ type AzureSQLDatabaseLinkedServiceTypeProperties struct {
 
 	// The name or ID of the tenant to which the service principal belongs. Type: string (or Expression with resultType string).
 	Tenant any
+
+	// Indicate whether the channel will be encrypted while bypassing walking the certificate chain to validate trust, used by
+	// recommended version. Type: Boolean.
+	TrustServerCertificate any
+
+	// The user name to be used when connecting to server. Type: string.
+	UserName any
 }
 
 // AzureSQLMILinkedService - Azure SQL Managed Instance linked service.
@@ -5089,26 +5365,116 @@ func (a *AzureSQLMILinkedService) GetLinkedService() *LinkedService {
 
 // AzureSQLMILinkedServiceTypeProperties - Azure SQL Managed Instance linked service properties.
 type AzureSQLMILinkedServiceTypeProperties struct {
-	// REQUIRED; The connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
-	ConnectionString any
-
 	// Sql always encrypted properties.
 	AlwaysEncryptedSettings *SQLAlwaysEncryptedProperties
+
+	// The application workload type when connecting to a server, used by recommended version. Possible values are ReadOnly and
+	// ReadWrite. Type: string.
+	ApplicationIntent any
+
+	// The type used for authentication. Type: string.
+	AuthenticationType *AzureSQLMIAuthenticationType
 
 	// Indicates the azure cloud type of the service principle auth. Allowed values are AzurePublic, AzureChina, AzureUsGovernment,
 	// AzureGermany. Default value is the data factory regions’ cloud type. Type:
 	// string (or Expression with resultType string).
 	AzureCloudType any
 
+	// The default wait time (in seconds) before terminating the attempt to execute a command and generating an error, used by
+	// recommended version. Type: integer.
+	CommandTimeout any
+
+	// The number of re-connections attempted after identifying that there was an idle connection failure, used by recommended
+	// version. This must be an integer between 0 and 255. Type: integer.
+	ConnectRetryCount any
+
+	// The amount of time (in seconds) between each re-connection attempt after identifying that there was an idle connection
+	// failure, used by recommended version. This must be an integer between 1 and 60.
+	// Type: integer.
+	ConnectRetryInterval any
+
+	// The length of time (in seconds) to wait for a connection to the server before terminating the attempt and generating an
+	// error, used by recommended version. Type: integer.
+	ConnectTimeout any
+
+	// The connection string, used by legacy version. Type: string, SecureString or AzureKeyVaultSecretReference. Type: string,
+	// SecureString or AzureKeyVaultSecretReference.
+	ConnectionString any
+
 	// The credential reference containing authentication information.
 	Credential *CredentialReference
+
+	// The name of the database, used by recommended version. Type: string.
+	Database any
+
+	// Indicate whether TLS encryption is required for all data sent between the client and server, used by recommended version.
+	// Type: string.
+	Encrypt any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// The name or address of the partner server to connect to if the primary server is down, used by recommended version. Type:
+	// string.
+	FailoverPartner any
+
+	// The host name to use when validating the server certificate for the connection. When not specified, the server name from
+	// the Data Source is used for certificate validation, used by recommended
+	// version. Type: string.
+	HostNameInCertificate any
+
+	// Indicate whether User ID and Password are specified in the connection (when false) or whether the current Windows account
+	// credentials are used for authentication (when true), used by recommended
+	// version. Type: Boolean.
+	IntegratedSecurity any
+
+	// The minimum time, in seconds, for the connection to live in the connection pool before being destroyed, used by recommended
+	// version. Type: integer.
+	LoadBalanceTimeout any
+
+	// The maximum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MaxPoolSize any
+
+	// The minimum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MinPoolSize any
+
+	// If your application is connecting to an AlwaysOn availability group (AG) on different subnets, setting MultiSubnetFailover=true
+	// provides faster detection of and connection to the (currently) active
+	// server, used by recommended version. Type: Boolean.
+	MultiSubnetFailover any
+
+	// When true, an application can maintain multiple active result sets (MARS). When false, an application must process or cancel
+	// all result sets from one batch before it can execute any other batch on
+	// that connection, used by recommended version. Type: Boolean.
+	MultipleActiveResultSets any
+
+	// The size in bytes of the network packets used to communicate with an instance of server, used by recommended version. Type:
+	// integer.
+	PacketSize any
+
 	// The Azure key vault secret reference of password in connection string.
 	Password *AzureKeyVaultSecretReference
+
+	// Indicate whether the connection will be pooled or explicitly opened every time that the connection is requested, used by
+	// recommended version. Type: Boolean.
+	Pooling any
+
+	// The name or network address of the instance of SQL Server to which to connect, used by recommended version. Type: string.
+	Server any
+
+	// The credential of the service principal object in Azure Active Directory. If servicePrincipalCredentialType is 'ServicePrincipalKey',
+	// servicePrincipalCredential can be SecureString or
+	// AzureKeyVaultSecretReference. If servicePrincipalCredentialType is 'ServicePrincipalCert', servicePrincipalCredential can
+	// only be AzureKeyVaultSecretReference.
+	ServicePrincipalCredential SecretBaseClassification
+
+	// The service principal credential type to use in Server-To-Server authentication. 'ServicePrincipalKey' for key/secret,
+	// 'ServicePrincipalCert' for certificate. Type: string (or Expression with
+	// resultType string).
+	ServicePrincipalCredentialType any
 
 	// The ID of the service principal used to authenticate against Azure SQL Managed Instance. Type: string (or Expression with
 	// resultType string).
@@ -5119,6 +5485,13 @@ type AzureSQLMILinkedServiceTypeProperties struct {
 
 	// The name or ID of the tenant to which the service principal belongs. Type: string (or Expression with resultType string).
 	Tenant any
+
+	// Indicate whether the channel will be encrypted while bypassing walking the certificate chain to validate trust, used by
+	// recommended version. Type: Boolean.
+	TrustServerCertificate any
+
+	// The user name to be used when connecting to server. Type: string.
+	UserName any
 }
 
 // AzureSQLMITableDataset - The Azure SQL Managed Instance dataset.
@@ -7923,7 +8296,7 @@ func (c *Credential) GetCredential() *Credential { return c }
 // CredentialListResponse - A list of credential resources.
 type CredentialListResponse struct {
 	// REQUIRED; List of credentials.
-	Value []*ManagedIdentityCredentialResource
+	Value []*CredentialResource
 
 	// The link to the next page of results, if any remaining results exist.
 	NextLink *string
@@ -7939,6 +8312,24 @@ type CredentialReference struct {
 
 	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
 	AdditionalProperties map[string]any
+}
+
+// CredentialResource - Credential resource type.
+type CredentialResource struct {
+	// REQUIRED; Properties of credentials.
+	Properties CredentialClassification
+
+	// READ-ONLY; Etag identifies change in the resource.
+	Etag *string
+
+	// READ-ONLY; The resource identifier.
+	ID *string
+
+	// READ-ONLY; The resource name.
+	Name *string
+
+	// READ-ONLY; The resource type.
+	Type *string
 }
 
 // CustomActivity - Custom activity type.
@@ -17204,9 +17595,6 @@ type ManagedIdentityCredential struct {
 
 	// Credential description.
 	Description *string
-
-	// Managed identity credential properties.
-	TypeProperties *ManagedIdentityTypeProperties
 }
 
 // GetCredential implements the CredentialClassification interface for type ManagedIdentityCredential.
@@ -17217,30 +17605,6 @@ func (m *ManagedIdentityCredential) GetCredential() *Credential {
 		Description:          m.Description,
 		Type:                 m.Type,
 	}
-}
-
-// ManagedIdentityCredentialResource - Credential resource type.
-type ManagedIdentityCredentialResource struct {
-	// REQUIRED; Managed Identity Credential properties.
-	Properties *ManagedIdentityCredential
-
-	// READ-ONLY; Etag identifies change in the resource.
-	Etag *string
-
-	// READ-ONLY; The resource identifier.
-	ID *string
-
-	// READ-ONLY; The resource name.
-	Name *string
-
-	// READ-ONLY; The resource type.
-	Type *string
-}
-
-// ManagedIdentityTypeProperties - Managed identity type properties.
-type ManagedIdentityTypeProperties struct {
-	// The resource id of user assigned managed identity
-	ResourceID *string
 }
 
 // ManagedIntegrationRuntime - Managed integration runtime, including managed elastic and managed dedicated integration runtimes.
@@ -23561,18 +23925,101 @@ func (s *SQLServerLinkedService) GetLinkedService() *LinkedService {
 
 // SQLServerLinkedServiceTypeProperties - SQL Server linked service properties.
 type SQLServerLinkedServiceTypeProperties struct {
-	// REQUIRED; The connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
-	ConnectionString any
-
 	// Sql always encrypted properties.
 	AlwaysEncryptedSettings *SQLAlwaysEncryptedProperties
+
+	// The application workload type when connecting to a server, used by recommended version. Possible values are ReadOnly and
+	// ReadWrite. Type: string.
+	ApplicationIntent any
+
+	// The type used for authentication. Type: string.
+	AuthenticationType *SQLServerAuthenticationType
+
+	// The default wait time (in seconds) before terminating the attempt to execute a command and generating an error, used by
+	// recommended version. Type: integer.
+	CommandTimeout any
+
+	// The number of re-connections attempted after identifying that there was an idle connection failure, used by recommended
+	// version. This must be an integer between 0 and 255. Type: integer.
+	ConnectRetryCount any
+
+	// The amount of time (in seconds) between each re-connection attempt after identifying that there was an idle connection
+	// failure, used by recommended version. This must be an integer between 1 and 60.
+	// Type: integer.
+	ConnectRetryInterval any
+
+	// The length of time (in seconds) to wait for a connection to the server before terminating the attempt and generating an
+	// error, used by recommended version. Type: integer.
+	ConnectTimeout any
+
+	// The connection string, used by legacy version. Type: string, SecureString or AzureKeyVaultSecretReference. Type: string,
+	// SecureString or AzureKeyVaultSecretReference.
+	ConnectionString any
+
+	// The name of the database, used by recommended version. Type: string.
+	Database any
+
+	// Indicate whether TLS encryption is required for all data sent between the client and server, used by recommended version.
+	// Type: string.
+	Encrypt any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// The name or address of the partner server to connect to if the primary server is down, used by recommended version. Type:
+	// string.
+	FailoverPartner any
+
+	// The host name to use when validating the server certificate for the connection. When not specified, the server name from
+	// the Data Source is used for certificate validation, used by recommended
+	// version. Type: string.
+	HostNameInCertificate any
+
+	// Indicate whether User ID and Password are specified in the connection (when false) or whether the current Windows account
+	// credentials are used for authentication (when true), used by recommended
+	// version. Type: Boolean.
+	IntegratedSecurity any
+
+	// The minimum time, in seconds, for the connection to live in the connection pool before being destroyed, used by recommended
+	// version. Type: integer.
+	LoadBalanceTimeout any
+
+	// The maximum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MaxPoolSize any
+
+	// The minimum number of connections allowed in the connection pool for this specific connection string, used by recommended
+	// version. Type: integer.
+	MinPoolSize any
+
+	// If your application is connecting to an AlwaysOn availability group (AG) on different subnets, setting MultiSubnetFailover=true
+	// provides faster detection of and connection to the (currently) active
+	// server, used by recommended version. Type: Boolean.
+	MultiSubnetFailover any
+
+	// When true, an application can maintain multiple active result sets (MARS). When false, an application must process or cancel
+	// all result sets from one batch before it can execute any other batch on
+	// that connection, used by recommended version. Type: Boolean.
+	MultipleActiveResultSets any
+
+	// The size in bytes of the network packets used to communicate with an instance of server, used by recommended version. Type:
+	// integer.
+	PacketSize any
+
 	// The on-premises Windows authentication password.
 	Password SecretBaseClassification
+
+	// Indicate whether the connection will be pooled or explicitly opened every time that the connection is requested, used by
+	// recommended version. Type: Boolean.
+	Pooling any
+
+	// The name or network address of the instance of SQL Server to which to connect, used by recommended version. Type: string.
+	Server any
+
+	// Indicate whether the channel will be encrypted while bypassing walking the certificate chain to validate trust, used by
+	// recommended version. Type: Boolean.
+	TrustServerCertificate any
 
 	// The on-premises Windows authentication user name. Type: string (or Expression with resultType string).
 	UserName any
@@ -31357,7 +31804,7 @@ type WebActivityTypeProperties struct {
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
 	// string (or Expression with resultType string).
-	Headers map[string]*string
+	Headers map[string]any
 
 	// List of linked services passed to web endpoint.
 	LinkedServices []*LinkedServiceReference
@@ -31516,7 +31963,7 @@ type WebHookActivityTypeProperties struct {
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
 	// string (or Expression with resultType string).
-	Headers map[string]*string
+	Headers map[string]any
 
 	// When set to true, statusCode, output and error in callback request body will be consumed by activity. The activity can
 	// be marked as failed by setting statusCode >= 400 in callback request. Default is